Pass in default values hash when creating a bundle using with_defaults. Note that create_bundle now uses with_defaults but simply passes in an empty default value hash.
Add support for the Bundle class to keep track of default values per bundle. Add helper method to apply default values to a given object if it responds to setters equivalent to the default value hash's key names (as name + "=" converted to a symbol). Add missing path getter to Bundle.

Add trigger property to command. Not sure if this is correct, but it is being used in commmand examples currently. Made equivalent changes on the Java side as well. Apply default values to new Command instances when they are created.

Apply default values to new instances of Snippet when they are created.