Now sort after each call to BundleEntry.addBundle() since that's the only place that could break sort order. We won't be doing tons of adds, so this shouldn't be a performance bottleneck.
Moved all logic for getting all commands, menus, and snippets for a give bundle entry into BundleEntry.

Modified BundleEntry.processMembers to implement the semantics of our bundle scope precedence system. We process only the highest precedence bundle scope that we have in the entry. We process all bundle references (zero or more) and only one non-ref bundle element.

Modify BundleProcessor interface to include the BundleEntry that is involved with the bundle processing and we now return a boolean from BundleProcessor.processBundle to allow implementations to stop processing, if desired.