Currently the :current_file and :current_project options aren't implemented, because the java code in CommandElement cannot safely access the state from the explorer or editor.
Sandip and I have discussed, and it seems like a refactoring is needed to make this work. But such a refactoring is beyond the scope of this change.

As a side change, I also fixed the output_type definition to match the spec: if you specify a string, it should treat it as a path to a file.