Remove isSnippet predicate as it wasn't working the way we wanted and it is not needed.
Remove logic for "executing" a snippet. That is now being handled by overrides in the SnippetElement

Re-apply refactoring of invokeBlockCommand that was lost during the last merge. Now wrapping the CommandContext in a Ruby Context instance so we can do Ruby things with that object. This in particular is used to access dynamically added properties in the CommandContext as if they were Ruby properties.