Fix issue "Find In Project reports sync errors"
This was happening after running commands like:

> rake rails:freeze:gems

It turns out - when the 'rake rails:freeze:gems' command is run it create vendor/rails folder structure but before it finishes, it also deletes some sub folder.

The fileDeleted call back from file watcher was coming in. However it was not being handled correctly. The fileDeleted passes back the rooPath (that is being watched) and sub path that is being deleted. The sub path was being ignored. Now it is handled by refreshing the parent resource (if any) to the depth of one.