Improvements to :show_as_tooltip.
Automatically hide the tooltip after 10 seconds. 
If the user clicks in the tooltip, it gets the focus.
This allows the user to copy the text. Once focused
the tooltip can be dismissed by typing Escape or
clicking elsewhere. Adjusted the message to reflect
the new behavior.

We should really be using the Eclipse like type
F2 to focus mechanism to make the tooltip fully
accesible.