Fix issue ":show_as_tool tooltip is shown offscreen"
Reason:

We were displaying the tool tip next to the caret position in the editor. However the display coordinates of the caret may not be in the visible area of user's display because of the scroll bar position of the editor's scroll bars.

Fix:

If the caret is in the client area of editor show the tooltip below it. If not check if the caret's Y offset is within the client area of the editor, then show the tooltip below the caret line but near the left margin. Else show the tool tip below the first visible line of the editor.