Remove debug version of webterm library and replace with the compressed release version.
Duplicated old index.html into a debug version. It has some handy stuff in it for debugging, but there's no point in actually loading that page since it pulls in more code that we want.

Remove all debugging code from index.html and reference the compressed release version of the webterm lib.