Fix issue "Switching into block editing mode changes the font size"
It turns out that the text editor implementation tries to match size of block selection mode font with the normal selection mode font. However it does this only when switching to the block selection mode and does not react to the normal mode font changes while it is in the block selection mode. We listen to the changes in normal mode font and when in block selection mode simply toggle to the normal selection mode and back to get the effect of font adjustment.