Fix issue "doctype HTML snippet should move cursor after inserted text"
When trying to get the selection after the insertion of the snippet proposal we need to ask for the selection from the snippet proposal that actually inserted the snippet. This can happen if they invoke the snippet proposal using the accelerator 1 thru 9 and the snippet proposal actually selected in the completion popup was  different than the one they selected using the accelerator.