Fix issues "Via Tender: Tabbing after "DIV" tag for autocomplete throws error."
The NPE was caused because we were attempting to return selection from the proposal that wraps a command and not a snippet. We now return null. The case of commands that insert the output as snippets is taken care of by the fact that we create a separate SnippetTemplateProposal to insert such output.