Fix issue "output_to_console in bundles does not do anything"
The output is going to the RadRails Scripting Console. The issue that the output is not visible when a dark background theme is in effect. That is because the foreground is not being used for normal output.

For CONSOLE_OUTPUT stream we should use the foreground color from the theme.