Fix issue "RadRails froze after switching project in windows"
The original reason for this was issue of passphrase prompt. The stack tarce showed that it was hanging on populating the branches combo. Max has now fixed the issue of passphrase prompt by inserting native shim which is executed instead of the actual ssh.exe through the use of GIT_SSH environment variable. However now we run into another issue. This current code in c.a.git.core tries to search and execute the .../bin/git.exe from msysgit installation folders. When it is run the first time to say - clone a github repository and after that proceeds as if there was no ssh key and fails with permission denied.

The likely diagnosis for this is that - when the git.exe run directly, the HOME variable may not be set correctly to point to the folder which contains the .ssh folder. That is why msysgit provides a wrapper in .../cmd/git.cmd which sets the HOME variable and then calls the git.exe. SO in our implementation we should prefer git.cmd over git.exe got the msysgit case.