Fix issue Studio 3 #285 "ad clicks open in external browser on windows"
It turns out that the event.required is not being set to true for the IE implementation of the SWT Browser widget on windows, when the hyperlink being clicked on specifies a "new" window in the target attribute. In any case it appears that the event callback happens correctly only when the browser has determined that it needs to create a new window. Thus we simply drop the check for event.required == true and always supply the new internal browser editor in the call back.