Take prefix into account when inserting values.
Check for element tokens when determining placement inside of a rule. This occurs while typing a property name. For example "b" appears to be an element while typing "background"

Make check for inside/outside rule a little more robust

Use the isSuggestedSelection property on CommonCompletionProposals when setting the currently selected item.