Add another recovery strategy to handle args in parentheses.
Use the "top" to access items in _symbols now that it is protected and no longer private.

Add comment about a way we could make selection of recovery strategies more efficient, which we may need to consider if we start getting a lot of recovery strategies.

Fix token ranges when fixing unfinished get-property statements.