Fix ROR3 #724 - Horizontal scroll trips
We weren't doing annotation updates properly for folding. We needed to handle when an edit shifted a number of annotations' positions some set offset. After a lot of work trying to be smart about diffing the new and old positions, I finally found a much simpler fix. Positions are updated in the model automatically and we don't need to mark the modified ones in any special way. Just iterate over all our old annotations, ask the model for their position and see if it's in the new set of positions. If it is, do nothing. If it isn't it's a new annotation. If we have some in the old set whose positions no longer exist, they've been deleted.