Attempt to fix AS3 #803 - Indexing error
try to be better about logging when the index is null for IndexContainer/IndexFilesofProjectJobs. Also try to make our code smarter and help avoid the situation where it could happen in the first place: When we're unable to re-use an existing index file because it blows up on initialize(), force generation of a new index file on disk that is empty save for the header/signature. Then schedule an immediate forced rebuild of the index that ignores timestamps (in case it was modified on a per file basis before it runs).