Implement https://www.pivotaltracker.com/story/show/4416784 in a way.
Add a \"track\" checkbox to the Add Remote dialog. When checked, it'll run \"git remote add --track current_branch remote_name uri\" instead of \"git remote add remote_name uri\". Try to default to having it on when the remote name is \"origin\". This should help users who are trying to hook up a local git repo to a new github repo.