Implement [Studio #1284] App Explorer does not allow user to see all files in project
The \".*\" filename filter is attached to our App Explorer by default and is set to active. unfortunately the extension point doesn't allow us to override the active by default setting. I added a hack that will turn it off once (basically the first time) programmatically. If the user turns it back on, we won't change that.