Fix [Studio #1388] - Keybinding to switch branches doesn't work
Actually what was happoening is that we'd get a return code of the index of the item chosen, and since we only did something if that value was 0, anything chosen other than the first item was a no-op. Now I check if the return value is != -1 which is used when cancelling.