Fix [Studio 3 #1464] - IOUtil.read deletes the last new line in the stream
Now IOUtil will read in the contents exactly as they are. Before it converted newlines into \"\n\" and also stripped trailing newlines from the input.

I took a look at all the callers and it seems like they should be ok in handling this change in behavior. Hopefully we won't see some odd newline issues as a result...