Fix for [Studio 3 #1913] - Batch calculation of pull indicators for same remote (all branches on that remote)
replace shouldPull(branchName) with getOutOfDateBranches(), which should be much more performant. Instead of asking if there are changes to the remote of a local branch for each and every branch, we batch and do it for all of them. This avoids ls-remote per remote branch, down to one ls-remote per remote (that has branches we care about).