Fix for #1997 - Line number color display inconsistent
There's a quick flash of fg/bg color being used as the line # background when you cahnge the theme or invasive option, but we then override and set it to gray again.

You may still get a non-grey bg to the line numbers if it's not our editor, just as a result of how eclipse uses the fg/bg colors.