Fix for LH2174. We have a preference (on by default) that allows users to turn off the "migration builders"
Implemented in the simplest way possible. This code could have been refactored to be better, but suggest for 3.1, we remove it all and create a ruble or menu command for it so people can run it on demand if necessary. It seemed unnecessary to make it fancy now if we're just going to get rid of it.