Work for [Studio 3 #2189] Via Tender: Git checkout from GUI (App Explorer) sometimes doesn't checkout branch
if we fail to switch branches, surface up the underlying error output in an error dialog or tooltip.

To enable this, I needed to introduce an IStatus subclass, ProcessStatus which held stdout and stderr separately and efault it's getMessage() to old behavior. Then in cases where we need to, we can cast to ProcessStatus and grab stdout/stderr explicitly.