Fix for #2406 - Pair matching causing delays in large JS files
Instead of checking scope everytime we encounter a character that might be part of a pair, we now just skip partitions wholesale if they contain \"_comment\" or \"_string\". This behavior is _not_identical as we rely on internal naming convetnions for partitions in this new manner, and in some cases partitions that were comments or string but didn't contain those magic strings will no longer be skipped for pair matching. However, this new way of handling pair matching is magnitudes faster as we not only avoid the scope check for each pair character found, but skip checking inside comment or string partitions altogether.