Add debug option messaging for file and resource monitoring, element registration, and bundle loading
Now check that parent directory of cache.yml is writable before trying to create a cache file. This avoids a potential exception we can get when a bundle is being deleted.

Now check that a script is readable before trying to run it through jruby. This prevents a potential exception that can occur when a bundle is being deleted.

Be a little more paranoid about exceptions occuring during firing of events in the bundle mananger. Previously, an exception could prevent reminaining listeners from being fired