Aptana views LinuxMint as an Ubuntu flavor
There was a bug in Aptana when run under Ubuntu that made it so Aptana drew it's own tree node images ON TOP OF the OS tree node images which looked really yucky.  Someone fixed this bug, but during OS detection, they forgot to include LinuxMint which is a sister operating system of Ubuntu, sharing most of it's OS qualities.  

https://jira.appcelerator.org/browse/APSTUD-2705?focusedCommentId=219421#comment-219421